GCC: Use explicit type casting for base::StringPrintf()
Use explicit type casting from net::ProxyResolveDnsOperation to int
for format ‘%d’ for base::StringPrintf() to avoid warning from GCC:
warning: format ‘%d’ expects argument of type ‘int’,
but argument 2 has type ‘net::ProxyResolveDnsOperation’

Bug: 819294
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3320391
Commit-Queue: Ivan Murashov <ivan.murashov@lge.com>
Cr-Commit-Position: refs/heads/main@{#949576}

+++ services/proxy_resolver/proxy_resolver_v8_tracing.cc
@@ -861,7 +861,7 @@ void Job::SaveDnsToLocalCache(const std::string& host,
 
 std::string Job::MakeDnsCacheKey(const std::string& host,
                                  net::ProxyResolveDnsOperation op) {
-  return base::StringPrintf("%d:%s", op, host.c_str());
+  return base::StringPrintf("%d:%s", static_cast<int>(op), host.c_str());
 }
 
 void Job::HandleAlertOrError(bool is_alert,
